Added a condition around the local file removal logic to only do it if the .desktop file exists somewhere other than the user's local directory structure.

+++ usr/lib/linuxmint/mintMenu/plugins/applications.py
@@ -750,11 +750,12 @@ class pluginclass( object ):
 
 		os.system('gnome-desktop-item-edit ' + file_path) 
 
-		if filecmp.cmp(widget.desktopFile, file_path):
-			try:
-				os.remove(file_path)
-			except os.error:
-				pass
+		if file_path != widget.desktopFile:
+			if filecmp.cmp(widget.desktopFile, file_path):
+				try:
+					os.remove(file_path)
+				except os.error:
+					pass
 
 		self.mintMenuWin.hide()
